+++ CHANGELOG.md
@@ -16,6 +16,7 @@ on the willingness of the author to come up with any and stick to it.
 
 * Tabs will now be rendered using 4 spaces, instead of showing up as `^I`. This
   will be made configurable in the future.
+* Uncommitted lines will no longer show up with a "000..." SHA.
 
 ### Fixed